DELIA-59539: [Coverity] Various medium impact issues identified in aamp - part 6
Reason for change: Coverity issues are addressed as requested from security team.
Test Procedure: Attached in the ticket
Risks: low

(cherry picked from commit 7e4764b8c023c7b41842c491a5ec46ace2092cfb)

+++ isobmff/isobmffbox.cpp
@@ -369,7 +369,7 @@ MvhdBox* MvhdBox::constructMvhdBox(uint32_t sz, uint8_t *ptr)
 {
 	uint8_t version = READ_VERSION(ptr);
 	uint32_t flags  = READ_FLAGS(ptr);
-	uint64_t tScale;
+	uint32_t tScale;
 
 	uint32_t skip = sizeof(uint32_t)*2;
 	if (1 == version)
